requirements.txt
shapely
numpy
scipy
pyyaml
